+++ CHANGELOG.md
@@ -7,6 +7,7 @@ List of changes that are finished but not yet released in any final version.
 - [PR-83](https://github.com/Knotx/knotx-fragments/pull/83) - Adding node log to inline body action.
 - [PR-82](https://github.com/Knotx/knotx-fragments/pull/82) - Adding logging to in-memory cache action.
 - [PR-60](https://github.com/Knotx/knotx-fragments/pull/60) - Adding node log to circuit breaker action. Enforce fallback on error.
+- [PR-45](https://github.com/Knotx/knotx-fragments/pull/46) - Fragment Event Consumer mechanism implementation
 
 ## Version 2.1.0
 - [PR-62](https://github.com/Knotx/knotx-fragments/pull/62) - It introduces task & graph node factories. It is connected with issue #49.